chore: refactor the refresh oauth services function (#31385)
