+++ apps/meteor/app/lib/server/methods/refreshOAuthService.ts
@@ -1,8 +1,7 @@
-import { Settings } from '@rocket.chat/models';
 import type { ServerMethods } from '@rocket.chat/ui-contexts';
 import { Meteor } from 'meteor/meteor';
-import { ServiceConfiguration } from 'meteor/service-configuration';
 
+import { refreshLoginServices } from '../../../../server/lib/refreshLoginServices';
 import { hasPermissionAsync } from '../../../authorization/server/functions/hasPermission';
 
 declare module '@rocket.chat/ui-contexts' {
@@ -29,8 +28,6 @@ Meteor.methods<ServerMethods>({
 			});
 		}
 
-		await ServiceConfiguration.configurations.removeAsync({});
-
-		await Settings.update({ _id: /^(Accounts_OAuth_|SAML_|CAS_).+/ }, { $set: { _updatedAt: new Date() } }, { multi: true });
+		await refreshLoginServices();
 	},
 });
