+++ apps/meteor/app/lib/server/startup/index.ts
@@ -1,10 +1,5 @@
-import { customOAuthServicesInit } from './oAuthServicesUpdate';
 import './rateLimiter';
 import './robots';
 import './settingsOnLoadCdnPrefix';
 import './settingsOnLoadDirectReply';
 import './settingsOnLoadSMTP';
-
-export const libStartup = async () => {
-	await customOAuthServicesInit();
-};
