+++ apps/meteor/ee/server/configuration/oauth.ts
@@ -21,8 +21,8 @@ interface IOAuthUserIdentity {
 }
 
 interface IOAuthSettings {
-	mapChannels: string;
-	mergeRoles: string;
+	mapChannels: boolean;
+	mergeRoles: boolean;
 	rolesToSync: string;
 	rolesClaim: string;
 	groupsClaim: string;
@@ -34,13 +34,13 @@ const logger = new Logger('EECustomOAuth');
 
 function getOAuthSettings(serviceName: string): IOAuthSettings {
 	return {
-		mapChannels: settings.get(`Accounts_OAuth_Custom-${serviceName}-map_channels`) as string,
-		mergeRoles: settings.get(`Accounts_OAuth_Custom-${serviceName}-merge_roles`) as string,
-		rolesToSync: settings.get(`Accounts_OAuth_Custom-${serviceName}-roles_to_sync`) as string,
-		rolesClaim: settings.get(`Accounts_OAuth_Custom-${serviceName}-roles_claim`) as string,
-		groupsClaim: settings.get(`Accounts_OAuth_Custom-${serviceName}-groups_claim`) as string,
-		channelsAdmin: settings.get(`Accounts_OAuth_Custom-${serviceName}-channels_admin`) as string,
-		channelsMap: settings.get(`Accounts_OAuth_Custom-${serviceName}-groups_channel_map`) as string,
+		mapChannels: settings.get<boolean>(`Accounts_OAuth_Custom-${serviceName}-map_channels`),
+		mergeRoles: settings.get<boolean>(`Accounts_OAuth_Custom-${serviceName}-merge_roles`),
+		rolesToSync: settings.get<string>(`Accounts_OAuth_Custom-${serviceName}-roles_to_sync`),
+		rolesClaim: settings.get<string>(`Accounts_OAuth_Custom-${serviceName}-roles_claim`),
+		groupsClaim: settings.get<string>(`Accounts_OAuth_Custom-${serviceName}-groups_claim`),
+		channelsAdmin: settings.get<string>(`Accounts_OAuth_Custom-${serviceName}-channels_admin`),
+		channelsMap: settings.get<string>(`Accounts_OAuth_Custom-${serviceName}-groups_channel_map`),
 	};
 }
