+++ apps/meteor/server/configuration/index.ts
@@ -1,0 +1,11 @@
+import { configureAccounts } from './accounts_meld';
+import { configureCAS } from './cas';
+import { configureLDAP } from './ldap';
+import { configureOAuth } from './oauth';
+
+export async function configureLoginServices() {
+	await configureAccounts();
+	await configureCAS();
+	await configureLDAP();
+	await configureOAuth();
+}
