+++ apps/meteor/server/lib/oauth/logger.ts
@@ -1,0 +1,3 @@
+import { Logger } from '@rocket.chat/logger';
+
+export const logger = new Logger('rocketchat:lib');
