+++ apps/meteor/server/main.ts
@@ -5,10 +5,11 @@ import './models/startup';
  * and the startup should be done in parallel
  */
 import './settings';
+import '../app/lib/server/startup';
 
-import { libStartup } from '../app/lib/server/startup';
 import { startLicense } from '../ee/app/license/server/startup';
 import { registerEEBroker } from '../ee/server';
+import { configureLoginServices } from './configuration';
 import { configureLogLevel } from './configureLogLevel';
 import { registerServices } from './services/startup';
 import { startup } from './startup';
@@ -23,17 +24,13 @@ await import('../lib/oauthRedirectUriServer');
 
 await import('./lib/pushConfig');
 
-await import('./configuration/accounts_meld');
-await import('./configuration/cas');
-await import('./configuration/ldap');
-
 await import('./stream/stdout');
 await import('./features/EmailInbox/index');
 
-await libStartup();
 await configureLogLevel();
 await registerServices();
 await import('../app/settings/server');
+await configureLoginServices();
 await registerEEBroker();
 await startup();
 await startLicense();
