+++ packages/ui-contexts/src/index.ts
@@ -1,5 +1,5 @@
 export { AttachmentContext, AttachmentContextValue } from './AttachmentContext';
-export { AuthenticationContext, AuthenticationContextValue, LoginService } from './AuthenticationContext';
+export { AuthenticationContextValue, AuthenticationContext, LoginService } from './AuthenticationContext';
 export { AuthorizationContext, AuthorizationContextValue } from './AuthorizationContext';
 export { AvatarUrlContext, AvatarUrlContextValue } from './AvatarUrlContext';
 export { ConnectionStatusContext, ConnectionStatusContextValue } from './ConnectionStatusContext';
